requirements.txt
requests==2.11.1
requests-toolbelt==0.7.0
moviepy==0.2.2.11
InstagramAPI==1.0.2
redis==2.10.5
celery==4.2.1
Pillow==5.4.1